fix: clean up unused directories and organize codebase structure
- Remove 8 empty directories (maintenance, migrations, services, utils, pages, store)
- Consolidate validation directory into single file
- Move misplaced test files to appropriate locations
- Consolidate test directories (tests -> __tests__)
- Document new structure in DIRECTORY_STRUCTURE.md
- Fix linting errors in validation.ts

Resolves #32
